bbpbat_backend_project/requirements.txt
asgiref==3.8.1
attrs==25.3.0
boto3==1.29.7
botocore==1.32.7
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.0
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==7.0.1
Django==4.2.7
django-allauth==0.57.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.10
granian==2.2.6
gunicorn==21.2.0
holidays==0.78
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
oauthlib==3.3.1
packaging==25.0
Pillow==10.0.1
psycopg2-binary==2.9.7
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.26.0
s3transfer==0.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.0.7
whitenoise==6.6.0
reportlab==4.0.0
pypdf==3.17.0
